requirements.txt
fastapi
uvicorn
cassandra-driver
python-dotenv
email-validator
